Revert all the Native Resource Code since that was not the cause of the failed signatures. Signatures failed because the EXE file was signed then changed which doesn't allow for resigning. Key is to use an unsigned EXE to build the package and then optionally resign the generated EXE.
There are still package format issues potentially that can get flagged by anti-virus and defender SmartScreen, but with signature this might be more secure.

+++ WebPackageViewer/FilePackager.cs
@@ -2,7 +2,8 @@ using System;
 using System.Diagnostics;
 using System.IO;
 using System.IO.Compression;
-using System.Reflection;
+using System.Text;
+using System.Windows;
 using WebPackageViewer.Utilities;
 
 namespace WebPackageViewer
@@ -12,13 +13,20 @@ namespace WebPackageViewer
     /// A file packager that can package a folder as a Zip file and 
     /// package into single file Exe and back into the 
     /// separate EXE and data parts.
-    /// 
-    /// It uses Native Resources that are embedded directly into the 
-    /// WebPackager exe and extracted from it.
     /// </summary>
     public class FilePackager
     {
-    
+        /// <summary>
+        /// String that separates the EXE binary and the data attached.
+        /// </summary>
+        public string SepararatorString = "--- DATASEPARATOR ---";
+
+        /// <summary>
+        /// Separator as a byte array
+        /// </summary>
+        public byte[] SeparatorBytes => Encoding.UTF8.GetBytes(SepararatorString ?? string.Empty);
+
+
         /// <summary>
         /// Packages a file by combining the package EXE file and the datafile.
         /// 
@@ -45,15 +53,21 @@ namespace WebPackageViewer
             if (File.Exists(packageFilename))           
                 File.Delete(packageFilename);
 
-            File.Copy(exeFilename, packageFilename, true);
-
-
-            // updates WEBSITE_DATA native resource with the zipped up Web site data 
-            // in byte form. 
-            NativeResourceHelper.UpdateResource(packageFilename, dataFilename, "WEBSITE_DATA");
-          
-
-            if (!string.IsNullOrEmpty(App.CommandLine.SignCommand))
+            using (var outFs = new FileStream(packageFilename, FileMode.Create, FileAccess.Write))
+            {
+                using (var fs = new FileStream(exeFilename, FileMode.Open, FileAccess.Read, FileShare.Read))
+                {
+                    fs.CopyTo(outFs);
+                    outFs.Flush();
+                    outFs.Write(SeparatorBytes, 0, SeparatorBytes.Length);
+                };
+                using (var fs = new FileStream(dataFilename, FileMode.Open, FileAccess.Read, FileShare.Read))
+                {
+                    fs.CopyTo(outFs);
+                }
+            }
+			
+			if (!string.IsNullOrEmpty(App.CommandLine.SignCommand))
             {
                 var cmd = App.CommandLine.SignCommand.Replace("%1", "\"" + packageFilename + "\"");
                 try
@@ -64,6 +78,7 @@ namespace WebPackageViewer
                 {
                 }
             }
+
             return true;
         }
 
@@ -130,22 +145,37 @@ namespace WebPackageViewer
             
             Directory.CreateDirectory(outputPath);
 
+            var offset = FindMarkerOffset(packageFilename, SeparatorBytes);
+            if (offset < 0)
+            {
+                SetError("File is missing Packaged Html Data.");
+                return false;
+            }
 
-            var exeFile = Path.Combine(outputPath, "WebPackageViewer.exe");
-
-
+            ;
             
+            var exeFile = Path.Combine(outputPath, "WebPackageViewer.exe");
             if (File.Exists(exeFile))
             {
                 File.Delete(exeFile);                
             }   
-        
-            File.Copy( Assembly.GetExecutingAssembly().Location, exeFile, true );
-                        
+            using (var exeFs = new FileStream(exeFile, FileMode.OpenOrCreate, FileAccess.Write, FileShare.Write))
+            using (var fs = new FileStream(packageFilename, FileMode.Open, FileAccess.Read, FileShare.Read))
+            {
+                CopyToBytes(fs, exeFs, offset - SeparatorBytes.Length);
+                fs.Flush();
+            }
+            
+
             var packageFile = Path.Combine(outputPath, "Packaged.zip");
-            var siteBytes = NativeResourceHelper.ReadResource("WEBSITE_DATA");            
-           
-            File.WriteAllBytes(packageFile, siteBytes);
+            
+            using (var packageFs = new FileStream(packageFile, FileMode.Create, FileAccess.Write, FileShare.Write))
+            using (var fs = new FileStream(packageFilename, FileMode.Open, FileAccess.Read, FileShare.Read))
+            {
+                fs.Seek(offset, SeekOrigin.Begin);
+                fs.CopyTo(packageFs);
+                packageFs.Flush();
+            }
             
             if (unZip)
             {                
@@ -155,11 +185,6 @@ namespace WebPackageViewer
                 }
                 //try { File.Delete(packageFile); } catch { /* ignore */ }
             }
-
-            if (App.CommandLine.RemoveResources)
-                NativeResourceHelper.UpdateResource(exeFile, "CLEAR", "WEBSITE_DATA");
-
-
 
             return true;
         }
@@ -225,6 +250,72 @@ namespace WebPackageViewer
             return true;
         }
 
+
+        /// <summary>
+        /// Like CopyTo but allows you to specify the number of bytes to copy and a buffer size.
+        /// </summary>
+        /// <param name="source"></param>
+        /// <param name="destination"></param>
+        /// <param name="bytesToCopy"></param>
+        /// <param name="bufferSize"></param>
+        void CopyToBytes(Stream source, Stream destination, long bytesToCopy, int bufferSize = 81920)
+        {
+            byte[] buffer = new byte[Math.Min(bufferSize, (int)bytesToCopy)];
+            int bytesRead;
+            long totalCopied = 0;
+
+            while (totalCopied < bytesToCopy &&
+                   (bytesRead = source.Read(buffer, 0, (int)Math.Min(buffer.Length, bytesToCopy - totalCopied))) > 0)
+            {
+                destination.Write(buffer, 0, bytesRead);
+                totalCopied += bytesRead;
+            }
+        }
+
+        /// <summary>
+        /// Find the offset of a marker in a file.
+        /// </summary>
+        /// <param name="exePath"></param>
+        /// <param name="marker"></param>
+        /// <returns></returns>
+        public long FindMarkerOffset(string exePath, byte[] marker)
+        {            
+            const int bufferSize = 4096;
+            byte[] buffer = new byte[bufferSize + marker.Length - 1];
+
+            using (var fs = new FileStream(exePath, FileMode.Open, FileAccess.Read))
+            {
+                long position = 0;
+                int bytesRead;
+
+                while ((bytesRead = fs.Read(buffer, 0, buffer.Length)) > 0)
+                {
+                    for (int i = 0; i <= bytesRead - marker.Length; i++)
+                    {
+                        bool match = true;
+                        for (int j = 0; j < marker.Length; j++)
+                        {
+                            if (buffer[i + j] != marker[j])
+                            {
+                                match = false;
+                                break;
+                            }
+                        }
+
+                        if (match)
+                            return position + i + marker.Length;
+                    }
+
+                    // Slide window back for overlap (in case marker splits across reads)
+                    if (bytesRead < marker.Length)
+                        break;
+                    position += bytesRead - marker.Length + 1;
+                    fs.Seek(position, SeekOrigin.Begin);
+                }
+            }
+
+            return -1;
+        }
 
         static void ExecuteCommandLine(string fullCommandLine,
             string workingFolder = null,
